+++ apps/tenant-management-webapp/src/app/lib/timeUtil.ts
@@ -1,0 +1,14 @@
+export const FormatDateTimeWithAt = (date: Date): string => {
+  const options: Intl.DateTimeFormatOptions = {
+    weekday: 'short',
+    month: 'short',
+    day: 'numeric',
+    hour: '2-digit',
+    minute: '2-digit',
+    year: 'numeric',
+  };
+
+  const timeParts = new Date(date).toLocaleString('en-US', options).split(',');
+  const localTime = `${timeParts[0]}, ${timeParts[1]}, ${timeParts[2]} at ${timeParts[3].toLowerCase()}`;
+  return localTime;
+};
